+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 // Providers
 import AuthProvider from './Providers/AuthProvider';
