+++ src/Components/Burger.tsx
@@ -7,12 +7,14 @@ const Burger = ({ open, setOpen, ...props }) => {
         type='button'
         className={open ? 'menu opened' : 'menu'}
         onClick={() => setOpen(!open)}
-        aria-label={open ? 'Close Menu' : 'Open Menu'}>
+        aria-label={open ? 'Close Menu' : 'Open Menu'}
+      >
         <svg
           width='50'
           height='50'
           viewBox='0 0 100 100'
-          aria-label='burger-menu'>
+          aria-label='burger-menu'
+        >
           <path
             className='line line1'
             d='M 20,29.000046 H 80.000231 C 80.000231,29.000046 94.498839,28.817352 94.532987,66.711331 94.543142,77.980673 90.966081,81.670246 85.259173,81.668997 79.552261,81.667751 75.000211,74.999942 75.000211,74.999942 L 25.000021,25.000058'
