+++ src/Components/Navbar.tsx
@@ -112,7 +112,8 @@ const Navbar = () => {
       <nav>
         <ul
           aria-hidden={open ? 'false' : 'true'}
-          className={open ? 'nav-links open' : 'nav-links'}>
+          className={open ? 'nav-links open' : 'nav-links'}
+        >
           {Pages.map((page) => {
             if (hasPermission(page.reqAuthState)) {
               return (
@@ -124,7 +125,8 @@ const Navbar = () => {
                     }`}
                     aria-hidden={open ? 'false' : 'true'}
                     tabIndex={open ? 0 : -1}
-                    onClick={handleLinkClick}>
+                    onClick={handleLinkClick}
+                  >
                     {page.name}
                   </Link>
                 </li>
@@ -141,7 +143,8 @@ const Navbar = () => {
                       type='button'
                       aria-hidden={open ? 'false' : 'true'}
                       tabIndex={open ? 0 : -1}
-                      onClick={() => button.action(handleLinkClick)}>
+                      onClick={() => button.action(handleLinkClick)}
+                    >
                       {button.label}
                     </button>
                   </div>
