+++ src/Pages/Home.tsx
@@ -13,8 +13,7 @@ const Home = () => {
           height: '40vh',
           width: '100%',
           textAlign: 'center',
-        }}
-      >
+        }}>
         <h1>&lt;/&gt; AvaRose</h1>
       </div>
       <section
@@ -22,8 +21,7 @@ const Home = () => {
           margin: '0 auto',
           textAlign: 'center',
           maxWidth: '800px',
-        }}
-      >
+        }}>
         <h2>🎉 Welcome to my website 👋</h2>
         <p>
           My name is Avery and I am a College Student at{' '}
