+++ src/interfaces/Card.ts
@@ -1,0 +1,5 @@
+export interface ICard {
+  id: number;
+  name: string;
+  body: string;
+}
