+++ src/interfaces/CardList.ts
@@ -1,0 +1,5 @@
+import { ICard } from './Card';
+
+export interface ICardList {
+  cards: ICard[];
+}
